Backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
dj-rest-auth==7.0.1
Django==5.1.7
django-allauth==65.6.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
oauthlib==3.2.2
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.1.0
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==78.1.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.3.0
wheel==0.45.1
